requirements.txt
requests==2.31.0
jinja2==3.1.3
python-dotenv==1.0.1
